+++ cli/cli/kurtosis_config/v1/kubernetes_cluster_config_v1.go
@@ -1,8 +1,0 @@
-package v1
-
-type KubernetesClusterConfigV1 struct {
-	KubernetesClusterName *string `yaml:"kubernetes-cluster-name,omitempty"`
-	StorageClass *string `yaml:"storage-class,omitempty"`
-	EnclaveSizeInMegabytes *uint `yaml:"enclave-size-in-Megabytes,omitempty"`
-}
-