+++ cli/cli/kurtosis_config/v1/kurtosis_cluster_config_v1.go
@@ -1,7 +1,0 @@
-package v1
-
-type KurtosisClusterConfigV1 struct {
-	Type *string                      `yaml:"type,omitempty"`
-	// If we ever get another type of cluster that has configuration, this will need to be polymorphically deserialized
-	Config *KubernetesClusterConfigV1 `yaml:"config,omitempty"`
-}