+++ cli/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+### Breaking Changes
+* Fixed bug in YAML spec where yaml type for enclave volume size ended in `Megabytes` not `megabytes`
 
 # 0.18.0
 ### Features