+++ cli/docs/changelog.md
@@ -1,6 +1,9 @@
 # TBD
 ### Features
-* Added extra guards to make it easier for Kurtosis devs to add new configuration versions
+* When the user adds a new config version, unit tests will now force (at compile time) the following:
+    * The user must add a new deserializer
+    * The user must add a new migration from the previous version
+    * The user must update `KurtosisConfig` to use the latest config struct
 
 ### Fixes
 * Renamed the `enclave-size-in-Megabytes` -> `enclave-size-in-megabytes` tag on the Kubernetes config